🩹 fix : Colors 타입 분리 및 추가
