+++ src/app/pokemons/_components/PokemonMoveColor.ts
@@ -1,23 +1,4 @@
-export type Colors = {
-  노말: string;
-  불꽃: string;
-  물: string;
-  풀: string;
-  전기: string;
-  얼음: string;
-  격투: string;
-  독: string;
-  땅: string;
-  비행: string;
-  에스퍼: string;
-  벌레: string;
-  바위: string;
-  고스트: string;
-  드래곤: string;
-  악: string;
-  강철: string;
-  페어리: string;
-};
+import type { Colors } from "@/types/Pokemon.type";
 
 export const typeColors: Colors = {
   노말: 'bg-zinc-400',
